refactor: signature authenticator
